backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv

# File Processing
PyPDF2
python-docx
python-multipart # For file uploads

# AI & Search
langchain           # Core LangChain
langchain-ollama    # Langchain Ollama integration

# Pydantic for models
pydantic[email]
pydantic-settings # For config loading

# Misc
requests # Might be used indirectly, keep for now
motor    # Async MongoDB driver
pymongo  # Standard MongoDB driver
passlib[bcrypt]
bcrypt==3.2.2 # Required for compatibility with passlib
PyJWT    # JWT tokens